[vim: plugins] lightline.vim configuration
Modified the information presented in the status line using
lightline.vim included functions. I don't use or find useful all the
information given in the stock implementation in lightline.vim, so I
decided to strip it down to the minimum needed information that I
personally use when writing code or prose.