[vim: color] Italics are out & quick-scope change
I commented out the option to use italics in jellybeans.vim colorscheme,
since Tamzyn doesn't have italic characters (maybe I should search for
an alternative bitmap font that has italics fonts to make use of this
option).

As of this commit, the quick-scope colours are not working with GUI
colors but are workin with 256 colors (at least inside of tmux I have no
problem with the highlighting of the shortcut characters), I could
switch completely to 256 colors or search for the bug that it's causing
me trouble with GUI colors.