[Vim: Tools] Added netwr to the tool belt
Discovered some time ago that netwr is an excellent built-in option to
NerdTree and file managing when in a buffer. I don't really need all the
funcionality from mature file tree plugins, so I decided to search for
some little hacks in the web and found this. I want to tinker with it in
more depth to add some stuff, but for now is good enough.