feat(lsp): Config neovim/nvim-lspconfig
Decided to give another shot at neovim/nvim-lspconfig since I updated
neovim. After all, my problems associated to being unable to config this
tool correctly was the incompatibility of my version of the package and
neovim. Now with that out of the way, some tinkering around and I was
able to reproduce the functionality I wanted in my setup.

Neovim rejoice!