feat: Change `neotex.nvim` for `vimtex.vim`
Decided to try a better out-of-shelf plugin for writing LaTeX or TeX
documents, since I'll proably write my whole undergrad thesis and future
work in this platform.

Have seen in a lot of places the power of vimtex as the standalone
plugin one must use in order to achieve seamless writing and editing
with the Vim experience.