feat: Changed "dracula" to "xresources"
Found out my lack of "magic" provided by `nvim-treesitter` was because
my version of `dracula.vim` didn't provided compatibility with given
plugin. Threfore I decided to change to `xresources` as it makes it
compatible with tree-sitter and it give me the flexibility if I want to
change my system colorscheme in a future.