✨ New colorschemes and 'tokyonight' as default
Decided to change from 'dracula' to 'tokyonight' as my default
colorscheme in vim since I liked the way it contrasted vim with the
terminal panes un 'tmux' and because I was never 100% happy with
'dracula'.

Will give 'tokyonight' a try since it's (1) lua based, therefore
extensible and (ii) comes with LOTS of integration with Lua based
plugins for Neovim.